+++ notepad_core/lib/NotepadClient.dart
@@ -9,6 +9,10 @@ import 'NotepadType.dart';
 import 'utils.dart';
 import 'woodemi/WoodemiClient.dart';
 
+abstract class NotepadClientCallback {
+  void handlePointer(List<NotePenPointer> list);
+}
+
 abstract class NotepadClient {
   static bool support(NotepadScanResult scanResult) {
     return startWith(scanResult.manufacturerData, WoodemiClient.prefix);
@@ -50,7 +54,14 @@ abstract class NotepadClient {
 
   NotepadType notepadType;
 
-  Future<void> completeConnection(void awaitConfirm(bool));
+  Future<void> completeConnection(void awaitConfirm(bool)) {
+    // TODO Cancel
+    notepadType.receiveSyncInput().listen((value) {
+      callback?.handlePointer(parseSyncData(value));
+    });
+  }
+
+  NotepadClientCallback callback;
 
   //#region authorization
   Uint8List _authToken;
@@ -82,5 +93,11 @@ abstract class NotepadClient {
   Future<int> getAutoLockTime(); // minute
 
   Future<void> setAutoLockTime(int time); // minute
+  //#endregion
+
+  //#region SyncInput
+  Future<void> setMode(NotepadMode notepadMode);
+
+  List<NotePenPointer> parseSyncData(Uint8List value);
   //#endregion
 }