+++ notepad_core/lib/models.dart
@@ -20,9 +20,37 @@ class NotepadScanResult {
       };
 }
 
+enum NotepadMode { Sync, Common }
+
 class BatteryInfo {
   final int percent;
   final bool charging;
 
   BatteryInfo(this.percent, this.charging);
+}
+
+class NotePenPointer {
+  int x;
+  int y;
+  int t;
+  int p;
+
+  NotePenPointer(this.x, this.y, this.t, this.p);
+
+  NotePenPointer.fromMap(map)
+      : this.x = map['x'],
+        this.y = map['y'],
+        this.t = map['t'],
+        this.p = map['p'];
+
+  Map toMap() => {
+    'x': x,
+    'y': y,
+    't': t,
+    'p': p,
+  };
+
+  @override
+  bool operator ==(other) =>
+      x == other.x && y == other.y && t == other.t && p == other.p;
 }