+++ notepad_core/CHANGELOG.md
@@ -1,13 +1,22 @@
+## [1.2.0] - 2019-12-21
+
+### NotepadConnector
+- Add `bluetoothChangeHandler` for bluetooth state change event
+
+### NotepadClient
+
+#### SyncInput
+- Add `setMode`
+- Add `NotepadClientCallback` with `handlePointer`
+
 ## [1.1.0] - 2019-12-20
 
 ### NotepadConnector
-
 - Add `connectionChangeHandler` for connection state change event
 
 ### NotepadClient
 
 #### Authorization
-
 - `connect` with authToken
 - `claimAuth`, `disclaimAuth`